refactor(core): promote pkg/headless to internal/core/headless (no behavior change)
