+++ internal/core/headless/headless.go
