+++ internal/core/headless/headless_test.go
